requirements.txt
pandas
numpy
pyyaml
streamlit